+++ resources/TankBase.png
